Hover over this to view changes.
More hovering compatibility.
Updated Settings to include ToolTips, hover over a setting to view more information to each setting.
Changed the outcome of the settings failing to load. You will not be able to change the option of the setting on startup.
Added Read-Only checking for the settings file. If it's read only, then the settings will not be writable, but still read-able.
Changed the LinkLabel to reflect on the new repo for youtube-dl-gui